add / edit, view and list project into components
