+++ app/imports/ui/pages/ListProject.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import { Meteor } from 'meteor/meteor';
-import { Button, Col, Container, Row } from 'react-bootstrap';
 import { useTracker } from 'meteor/react-meteor-data';
-import { Link } from 'react-router-dom';
 import { Projects } from '../../api/project/Projects';
-import Project from '../components/Project';
+import ProjectList from '../components/ProjectList';
 import LoadingSpinner from '../components/LoadingSpinner';
-import * as CTPNav from '../../api/navigation/Navigation';
 
 /* Renders a table containing all of the Stuff documents. Use <StuffItem> to render each row. */
 const ListProject = () => {
@@ -27,22 +24,7 @@ const ListProject = () => {
   }, []);
 
   return (ready ? (
-    <Container className="py-3">
-      <Row className="justify-content-center">
-        <Col md={7}>
-          <Col className="text-center">
-            <Row>
-              <Col> </Col>
-              <Col><h2>Projects</h2></Col>
-              <Col><Link to={CTPNav.PathAddProject}><Button variant="primary">Add Project</Button></Link></Col>
-            </Row>
-          </Col>
-          <Row>
-            {projects.map((project) => (<Row className="p-3" key={project._id}><Project project={project} /></Row>))}
-          </Row>
-        </Col>
-      </Row>
-    </Container>
+    <ProjectList projects={projects} />
   ) : <LoadingSpinner />);
 };
